% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add.PS.R
\name{Add.PS}
\alias{Add.PS}
\title{Add estimated propensity score to a data frame}
\usage{
Add.PS(data, trt = "D", ps.covs = c("x1", "x2", "x3", "x4"))
}
\arguments{
\item{data}{the name of a data frame containing all needed variables, including D (indicator of current trial) and a list of baseline covariates included in PS model}

\item{trt}{a character of variables name for the binary outcome in PS model. Defaul to "D"}

\item{ps.covs}{a list of characters for variable na,es included in PS model}
}
\description{
This function allows you to add estimated PS from propensity model logit(P("trt"=1)) = ps.covs
}
\details{
Returns data.ps, the data frame with one added column of est.PS for estimated PS.

Returns PS.fit, the glm fit of the propensity model
}
\author{
Xi "Ada" Wang

Ph.D. Student of Biostatistics

Penn State College of Medicine

xzw149@psu.edu
}
\keyword{AugRCT}
