% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extract.Hybrid.R
\name{Extract.Hybrid}
\alias{Extract.Hybrid}
\title{Extract the parameter of interest from a list of fits of Bayesian models (hybrid approaches)}
\usage{
Extract.Hybrid(fits, labels, par = "delta")
}
\arguments{
\item{fits}{a vector of rstan fits}

\item{labels}{a vector of character labels for fits of hybrid approaches}

\item{par}{the parameter of interest you want to extract from the fits. Default to "delta".}
}
\value{
res is the dataset with extracted estimates from fits
}
\description{
This function allows you to exract posterior estimates of interest from a list of fits of Bayesian models (hybrid approaches)
}
\details{
Returns res, a data frame of the point estimate and percentiles of a parameter of interest ( default is response rate difference between experiment and
augmented control group) of all the hybrid methods.
}
\author{
Xi "Ada" Wang

Ph.D. Student of Biostatistics

Penn State College of Medicine

xzw149@psu.edu
}
\keyword{AugRCT}
