% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Find.True.Simu.Setting.R
\name{find.lambda}
\alias{find.lambda}
\title{Find the true value of lambda such that desired power is achieved in the standard setting via simulation}
\usage{
find.lambda(
  n.sim = 1000,
  power = 0.8,
  seed.init = 2020,
  search.lowbound = 1,
  search.uppbound = 1.3
)
}
\arguments{
\item{n.sim}{number of simulation for power calculation}

\item{power}{desired power. Default to 0.8}

\item{seed.init}{seed for random number generator}

\item{search.lowbound}{lower bound to search the true value of lambda. Default to 1.}

\item{search.uppbound}{upper bound to search the true value of lambda. Default to 1.3.}
}
\description{
The value of lambda is found under the standard RCT 1:1 with 80 patients in experimental arm and 80 in control arm
}
\details{
Return the value of lambda
}
