% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Find.True.Simu.Setting.R
\name{find.xi}
\alias{find.xi}
\title{Find the true value of xi such that desired drift is achieved via simulation}
\usage{
find.xi(
  thetaCH,
  seed = 2020,
  lambda = 1.185,
  search.lowbound = -10,
  search.uppbound = 10
)
}
\arguments{
\item{thetaCH}{desired thetaCH.}

\item{seed}{seed for random number generator}

\item{lambda}{log odds of treatment effect. HR=exp(lambda). Default to 1.185, the scenario with a positive treatment effect s.t. a power of 0.8 is achived in RCT 1:1 (80 patients in E group, 80 in C group)}

\item{search.lowbound}{lower bound to search the true value of xi. Default to -10.}

\item{search.uppbound}{upper bound to search the true value of xi. Default to 10.}
}
\description{
The value of xi is found for desired value of drift by specifying thetaCH with different treatment effect (lambda)
}
\details{
Return the value of xi
}
