% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.data.R
\name{generate.data}
\alias{generate.data}
\title{Generate data for simulation}
\usage{
generate.data(
  n.E = 80,
  n.CD = 40,
  n.CH = 300,
  x1.mean.ECD = 40,
  x1.mean.CH = 40,
  x1.sd = 5,
  x2.p.ECD = 0.4,
  x2.p.CH = 0.4,
  x3.sd = 1,
  lambda = 1.19,
  beta0 = 3.6,
  beta1 = -0.1,
  beta2 = -0.5,
  beta3 = 1,
  beta4 = -1,
  xi = 0,
  seed = 2020
)
}
\arguments{
\item{n.E}{a positive integer for number of patients in current experimental arm (E). Default to 80.}

\item{n.CD}{a positive integer for number of patients in current control arm (CD). Default to 40.}

\item{n.CH}{a positive integer for number of patients in historical control arm (CH). Default to 300.}

\item{x1.mean.ECD}{mean of continuous variable X1 in the current trial. Default to 40}

\item{x1.mean.CH}{mean of continuous variable X1 in the historical trial. Default to 40}

\item{x1.sd}{standard deviation of continuous variable X1 in both the current and historical trials. Default to 5.}

\item{x2.p.ECD}{mean of binary variable X2 in the current trial. Default to 0.4}

\item{x2.p.CH}{mean of binary variable X2 in the historical trial. Default to 0.4}

\item{x3.sd}{standard deviation of continuous variable X3 in both the current and historical trials. Default to 1.}

\item{lambda}{log odds of treatment effect. HR=exp(lambda).}

\item{beta0}{intercept in the logistic outcome model. Default to 3.6}

\item{beta1}{coefficient of X1 in the logistic outcome model. Default to -0.1}

\item{beta2}{coefficient of X2 in the logistic outcome model. Default to -0.5}

\item{beta3}{coefficient of X3 in the logistic outcome model. Default to 1}

\item{beta4}{coefficient of X4 in the logistic outcome model. Default to -1}

\item{xi}{log odds of current study effect. = 0 if no drift.}

\item{seed}{a single value, interpreted as an integer, or NULL. seed of random number generator. Default to 2020}
}
\description{
This function allows you to generate data for simulation
}
\details{
Returns mydata, a data frame of simulated dataset, including ID, Z (indicator of treatment), D (indicator of current trial), group ("E", "CD", "CH"), x1, x2, x3, x4, Y (binary outcome)
}
\author{
Xi "Ada" Wang

Ph.D. Student of Biostatistics

Penn State College of Medicine

xzw149@psu.edu
}
\keyword{AugRCT}
