% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.iptw.ci.R
\name{get.iptw.ci}
\alias{get.iptw.ci}
\title{Calculate the point estimate and 95\% Bootstrap CI with IPTW approach}
\usage{
get.iptw.ci(data, label, alpha = 0.05)
}
\arguments{
\item{data}{the name of a data frame containing all needed variables, including Z (indicator of treatment) and Y (binary outcome).}

\item{label}{a character of labels for a hybrid approach}

\item{alpha}{level of Type I error. Default to 0.05 to require 95\% CI.}
}
\description{
This function allows you to calculate the point estimate and 95\% Bootstrap CI with IPTW approach of treatment effect in RD using IPTW1 estimator in Lunceford's paper (Stratification and Weighting Via the Propensity Score in Estimation of Causal Treatment Effects: A Comparative Study)
}
\details{
Returns a data frame with one row, and four columns for point.est, lower.CI, upper.CI and label.
augmented control group) of all the frequentist methods.
}
\author{
Xi "Ada" Wang

Ph.D. Student of Biostatistics

Penn State College of Medicine

xzw149@psu.edu
}
\keyword{AugRCT}
