% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.mn.ci.R
\name{get.mn.ci}
\alias{get.mn.ci}
\title{Calculate the point estimate and 95\% MN CI}
\usage{
get.mn.ci(data, label, alpha = 0.05)
}
\arguments{
\item{data}{the name of a data frame containing all needed variables, including Z (indicator of treatment) and Y (binary outcome).}

\item{label}{a character of labels for a hybrid approach}

\item{alpha}{level of Type I error. Default to 0.05 to require 95\% CI.}
}
\description{
This function allows you to calculate the point estimate and 95\% Miettinen and Nurminen CI of treatment effect in RD
}
\details{
Returns a data frame with one row, and four columns for point.est, lower.CI, upper.CI and label.
}
\author{
Xi "Ada" Wang

Ph.D. Student of Biostatistics

Penn State College of Medicine

xzw149@psu.edu
}
\keyword{AugRCT}
