% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.simu.R
\name{plot.simu}
\alias{plot.simu}
\title{Generate GGplot for methods comparison in simulation}
\usage{
plot.simu(
  fig.dt,
  fig.title,
  quants,
  ref.quants,
  ref.locs,
  ref.labels,
  cols,
  facet.row,
  facet.col,
 
    fig.path = "C:\\\\Users\\\\wangxi8\\\\OneDrive - Merck Sharp & Dohme, Corp\\\\AugRCT_new\\\\Simulation Results\\\\Figure\\\\",
  fig.width = 12,
  fig.height = 9,
  log_10 = FALSE
)
}
\arguments{
\item{fig.dt}{a data frame of reaults for figure}

\item{fig.title}{a character as the title of figure}

\item{quants}{a string vector of quantities to plot}

\item{ref.quants}{a string vector of quantities to add refrence lines}

\item{ref.locs}{a real number vector of locations for reference lines}

\item{ref.labels}{a vector of refline lables}

\item{cols}{a vector of color codes for plots}

\item{facet.row}{a string vector of variable names in rows for facet}

\item{facet.col}{a string vector of variable names in columns for facet}

\item{fig.path}{path to save the figure}

\item{fig.width}{width of figure}

\item{fig.height}{height of figure}

\item{log_10}{log10 transformaiton of y axis (used to enlarge difference of Type I error)}
}
\description{
This function generates a figure for comparison in simulation and save it under the specified path
}
\details{
Returns the ggplot subject for print and save it as .png under specified path
}
\author{
Xi "Ada" Wang

Ph.D. Student of Biostatistics

Penn State College of Medicine

xzw149@psu.edu
}
