% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RealData.Freq.R
\name{RealData.Freq}
\alias{RealData.Freq}
\title{Implement Frequentist (naive and PS only) approaches in real data application}
\usage{
RealData.Freq(
  realdata,
  ps.covs,
  nstrata = 5,
  trt = "D",
  seed = 2020,
  bootstrap.seed = 1234,
  bootstrap.rep = 10000
)
}
\arguments{
\item{realdata}{the name of a data frame containing all needed variables, including D (indicator of current trial), Z (indicator of treatment),
Y (binary outcome), a list of baseline covariates included in PS model}

\item{ps.covs}{a list of characters for variable na,es included in PS model. e.g. ps.covs=c("X1","X2")}

\item{nstrata}{a positive integer of number of strata in PS stratification. Default to 5.}

\item{trt}{a character of variables name for the binary outcome in PS model. Defaul to "D"}

\item{seed}{a single value, interpreted as an integer, or NULL. seed of random number generator. Default to 2020}

\item{bootstrap.seed}{seed for bootstrap CI of IPTW approach. Default to 1234}

\item{bootstrap.rep}{replication number of Bootstrap samples. Default to 10000}
}
\value{
comp.freq, a data frame of the point estimate and percentiles of delta (response rate difference between experiment and augmented control group) of all the frequentist methods.

pm.realdata, a data frame with PS matching adjusted patients.

realdata.iptw, a data frame with IPTW adjusted patients.

strat.trim.realdata, a data frame with PS stratification adjusted patients. variable strata in the data frame
indicates stratum number of each subject.
}
\description{
This function allows you to implement all the frequentist approaches (naive and PS only) in augmenting current control with historical data in real data application.
}
\author{
Xi "Ada" Wang

Ph.D. Student of Biostatistics

Penn State College of Medicine

xzw149@psu.edu
}
\keyword{AugRCT}
