% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RealData.Hybrid.R
\name{RealData.Hybrid}
\alias{RealData.Hybrid}
\title{Implement Hybrid approaches in real data application}
\usage{
RealData.Hybrid(
  realdata,
  ps.covs,
  trt = "D",
  nstrata = 5,
  m2.iter = 60000,
  m2.adapt_delta = 0.999,
  m2.max_treedepth = 12,
  m4.iter = 20000,
  m4.adapt_delta = 0.99,
  m5.iter = 20000,
  m5.adapt_delta = 0.99,
  seed = 2020
)
}
\arguments{
\item{realdata}{the name of a data frame containing all needed variables, including D (indicator of current trial), Z (indicator of treatment),
Y (binary outcome), a list of baseline covariates included in PS model}

\item{ps.covs}{a list of characters for variable na,es included in PS model}

\item{trt}{a character of variables name for the binary outcome in PS model. Defaul to "D"}

\item{nstrata}{a positive integer of number of strata in PS stratification. Default to 5.}

\item{m2.iter}{a positive integer of iteration number in STAN model 2 (Matching + Fixed power prior. Beta model). Default to 60000}

\item{m2.adapt_delta}{a real number between 0 and 1 to control the step size in STAN model 2. Default to 0.999}

\item{m2.max_treedepth}{a positive integer of tree depth of STAN model 2. Default to 12}

\item{m4.iter}{a positive integer of iteration number in STAN model 4 (Matching + Commensurate prior). Default to 20000}

\item{m4.adapt_delta}{a real number between 0 and 1 to control the step size in STAN model 4. Default to 0.99}

\item{m5.iter}{a positive integer of iteration number in STAN model 5 (PS stratification + Commensurate prior). Default to 20000}

\item{m5.adapt_delta}{a real number between 0 and 1 to control the step size in STAN model 5. Default to 0.99}

\item{seed}{a single value, interpreted as an integer, or NULL. seed of random number. Default to 2020}
}
\value{
comp, a data frame of the point estimate and percentiles of delta (response rate difference between experiment and augmented control group) of all the hybrid methods.

comp.fits, the list of the model fits for all the hybrid methods.
}
\description{
This function allows you to implement hybrid approaches (PS plus Bayesian) in augmenting current control with historical data in real data application.
}
\author{
Xi "Ada" Wang

Ph.D. Student of Biostatistics

Penn State College of Medicine

xzw149@psu.edu
}
\keyword{AugRCT}
