% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwe_functions.R
\name{rwePSDist}
\alias{rwePSDist}
\title{Get number of subjects and the distances of PS distributions for each PS strata}
\usage{
rwePSDist(data.withps, n.bins = 10, type = c("ovl", "kl"), ...)
}
\arguments{
\item{data.withps}{data frame with estimated PS}

\item{n.bins}{number of bins for KL computation. Default to 10}
}
\description{
Get number of subjects and the distances of PS distributions for each PS strata
}
