% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwe_functions.R
\name{rwePsPreprocess}
\alias{rwePsPreprocess}
\title{Get Posterior for all stratum}
\usage{
rwePsPreprocess(data, A = 0, RS = NULL, v.outcome = "Y", group = "group")
}
\arguments{
\item{data}{class DWITHPS data frame}

\item{A}{target number of subjects to be borrowed}

\item{RS}{parameters for dirichelet prior}

\item{type}{type of outcomes}

\item{...}{extra parameters for calling function \code{\link{rweSTAN}}}
}
\description{
Get Posterior for all stratum
}
