% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwe_functions.R
\name{rweSTAN}
\alias{rweSTAN}
\title{Run stan model}
\usage{
rweSTAN(
  lst.data,
  model,
  chains = 4,
  iter = 10000,
  warmup = 1000,
  control = list(adapt_delta = 0.95),
  ...
)
}
\description{
@param lst.data list of data to run STAN model
 @param model name of STAN model
 @export
}
