% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.freq.res.R
\name{save.freq.res}
\alias{save.freq.res}
\title{Save simulation results from frequentist methods (PS only methods)}
\usage{
save.freq.res(simu.server.path, filename, res.path, nsim)
}
\arguments{
\item{simu.server.path}{path of simulation results on server.}

\item{filename}{name of file for summarized simulation results, ended with ".Rdata". For example, "freq_5000.RData".}

\item{res.path}{path to save summarized simulation results, ended with "\\" or "/".}

\item{nsim}{simulation number.}
}
\description{
Save simulation results from frequentist methods (PS only methods)
}
