% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simu.Freq.R
\name{Simu.Freq}
\alias{Simu.Freq}
\title{Implement Frequentist (naive and PS only) approaches in simulation}
\usage{
Simu.Freq(
  seed,
  set.num = 0,
  n.E = 80,
  n.CD = 40,
  n.CH = 300,
  lambda = 2,
  xi = 0,
  x1.mean.ECD = 40,
  x1.mean.CH = 40,
  x2.p.ECD = 0.4,
  x2.p.CH = 0.4,
  beta0 = 3.6,
  beta1 = -0.1,
  beta2 = -0.5,
  beta3 = 1,
  beta4 = -1,
  nstrata = 3,
  trt = "D",
  ps.covs = c("x1", "x2", "x3", "x4"),
  bootstrap.seed = 1234,
  bootstrap.rep = 5000
)
}
\arguments{
\item{seed}{a single value, interpreted as an integer, or NULL. seed of random number generator.}

\item{set.num}{number of simulation setting, range from 1 to 32.}

\item{n.E}{a positive integer for number of patients in current experimental arm (E). Default to 80.}

\item{n.CD}{a positive integer for number of patients in current control arm (CD). Default to 40.}

\item{n.CH}{a positive integer for number of patients in historical control arm (CH). Default to 300.}

\item{lambda}{log odds of treatment effect. HR=exp(lambda).}

\item{xi}{log odds of current study effect. = 0 if no drift.}

\item{x1.mean.ECD}{mean of continuous variable X1 in the current trial. Default to 40}

\item{x1.mean.CH}{mean of continuous variable X1 in the historical trial. Default to 40}

\item{x2.p.ECD}{mean of binary variable X2 in the current trial. Default to 0.4}

\item{x2.p.CH}{mean of binary variable X2 in the historical trial. Default to 0.4}

\item{beta0}{intercept in the logistic outcome model. Default to 3.6}

\item{beta1}{coefficient of X1 in the logistic outcome model. Default to -0.1}

\item{beta2}{coefficient of X2 in the logistic outcome model. Default to -0.5}

\item{beta3}{coefficient of X3 in the logistic outcome model. Default to 1}

\item{beta4}{coefficient of X4 in the logistic outcome model. Default to -1}

\item{nstrata}{a positive integer of number of strata in PS stratification. Default to 3.}

\item{trt}{a character of variables name for the binary outcome in PS model. Defaul to "D"}

\item{ps.covs}{a list of characters for variable na,es included in PS model. Default to \code{c("x1","x2","x3","x4")}}

\item{bootstrap.seed}{seed for bootstrap CI of IPTW approach. Default to 1234}

\item{bootstrap.rep}{replication number of Bootstrap samples. Default to 5000}
}
\value{
mydata, a data frame of simulated dataset

comp.freq, a data frame of the point estimate and percentiles of delta (response rate difference between experiment and augmented control group) of all the frequentist methods.
}
\description{
This function allows you to implement all the frequentist approaches (naive and PS only) in augmenting current control with historical data in simulation.
}
\author{
Xi "Ada" Wang

Ph.D. Student of Biostatistics

Penn State College of Medicine

xzw149@psu.edu
}
\keyword{AugRCT}
