% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simu.Hybrid.R
\name{Simu.Hybrid}
\alias{Simu.Hybrid}
\title{Implement Hybrid approaches in simulation}
\usage{
Simu.Hybrid(
  seed,
  set.num = 0,
  n.E = 80,
  n.CD = 40,
  n.CH = 300,
  lambda = 2,
  xi = 0,
  x1.mean.ECD = 40,
  x1.mean.CH = 40,
  x2.p.ECD = 0.4,
  x2.p.CH = 0.4,
  beta0 = 3.6,
  beta1 = -0.1,
  beta2 = -0.5,
  beta3 = 1,
  beta4 = -1,
  nstrata = 3,
  m2.iter = 40000,
  m2.adapt_delta = 0.99,
  m2.max_treedepth = 12,
  m4.iter = 20000,
  m4.adapt_delta = 0.99,
  m5.iter = 20000,
  m5.adapt_delta = 0.99,
  trt = "D",
  ps.covs = c("x1", "x2", "x3", "x4")
)
}
\arguments{
\item{seed}{a single value, interpreted as an integer, or NULL. seed of random number. Default to 2020}

\item{set.num}{number of simulation setting, range from 1 to 32.}

\item{n.E}{a positive integer for number of patients in current experimental arm (E). Default to 80.}

\item{n.CD}{a positive integer for number of patients in current control arm (CD). Default to 40.}

\item{n.CH}{a positive integer for number of patients in historical control arm (CH). Default to 300.}

\item{lambda}{log odds of treatment effect. HR=exp(lambda).}

\item{xi}{log odds of current study effect. = 0 if no drift.}

\item{x1.mean.ECD}{mean of continuous variable X1 in the current trial. Default to 40}

\item{x1.mean.CH}{mean of continuous variable X1 in the historical trial. Default to 40}

\item{x2.p.ECD}{mean of binary variable X2 in the current trial. Default to 0.4}

\item{x2.p.CH}{mean of binary variable X2 in the historical trial. Default to 0.4}

\item{beta0}{intercept in the logistic outcome model. Default to 3.6}

\item{beta1}{coefficient of X1 in the logistic outcome model. Default to -0.1}

\item{beta2}{coefficient of X2 in the logistic outcome model. Default to -0.5}

\item{beta3}{coefficient of X3 in the logistic outcome model. Default to 1}

\item{beta4}{coefficient of X4 in the logistic outcome model. Default to -1}

\item{nstrata}{a positive integer of number of strata in PS stratification. Default to 3.}

\item{m2.iter}{a positive integer of iteration number in STAN model 2 (Matching + Fixed power prior. Beta model). Default to 40000}

\item{m2.adapt_delta}{a real number between 0 and 1 to control the step size in STAN model 2. Default to 0.99}

\item{m2.max_treedepth}{a positive integer of tree depth of STAN model 2. Default to 12}

\item{m4.iter}{a positive integer of iteration number in STAN model 4 (Matching + Commensurate prior). Default to 20000}

\item{m4.adapt_delta}{a real number between 0 and 1 to control the step size in STAN model 4. Default to 0.99}

\item{m5.iter}{a positive integer of iteration number in STAN model 5 (PS stratification + Commensurate prior). Default to 20000}

\item{m5.adapt_delta}{a real number between 0 and 1 to control the step size in STAN model 5. Default to 0.99}

\item{trt}{a character of variables name for the binary outcome in PS model. Defaul to "D"}

\item{ps.covs}{a list of characters for variable na,es included in PS model. Default to \code{c("x1","x2","x3","x4")}}
}
\value{
mydata, a data frame of simulated dataset

comp, a data frame of the point estimate and percentiles of delta (response rate difference between experiment and augmented control group) of all the hybrid methods.
}
\description{
This function allows you to implement hybrid approaches (PS plus Bayesian) in augmenting current control with historical data in simulation.
}
\details{
There might be warnings on the divergence of the STAN model fit. But the chains of posterior samples of the parameter of interest mixes well by checking the trace plot.
}
\author{
Xi "Ada" Wang

Ph.D. Student of Biostatistics

Penn State College of Medicine

xzw149@psu.edu
}
\keyword{AugRCT}
